requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
redis==5.0.8
boto3==1.35.23
python-dotenv==1.0.1
httpx==0.27.2
tenacity==9.0.0
loguru==0.7.2
openai==1.44.1
langgraph==0.2.38
langchain==0.3.4